Moved Config files to the root of includes/srcs to fix CI issue
